requirements.txt
anyio==4.4.0
apiservice @ file:///Users/user/Downloads/scaleit-api-main%20copy/apiservice-0.0.8-py3-none-any.whl#sha256=54a38a646eae9cf5581b7a36a1231950496a4098399f29fcb3c02ab7e4c98c07
assertpy==1.1
attrs==23.2.0
autohooks==24.2.0
autohooks-plugin-black==23.10.0
bcrypt==4.1.3
black==24.4.2
boto3==1.34.145
botocore==1.34.145
cachetools==5.4.0
Cerberus==1.3.5
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorful==0.5.6
cryptography==43.0.0
DateTime==4.9
decorator==5.1.1
deepdiff==7.0.1
distlib==0.3.8
et-xmlfile==1.1.0
exrex==0.11.0
filelock==3.15.4
future==1.0.0
gcloud==0.18.3
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpath==0.82.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
logger==1.4
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mock==5.1.0
mypy-extensions==1.0.0
numpy==2.0.1
oauth2client==4.1.3
openpyxl==3.1.5
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
paramiko==3.4.0
pathspec==0.12.1
pipenv==2024.0.1
platformdirs==4.2.2
pluggy==1.5.0
pontos==24.3.2
proto-plus==1.24.0
protobuf==5.28.0rc1
psycopg2-binary==2.9.9
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.3.1
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
retry==0.9.2
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
s3transfer==0.10.2
selenium==4.23.0
semver==3.0.2
shtab==1.7.1
singledispatch==4.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sshtunnel==0.4.0
testrail==0.3.15
tomlkit==0.13.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
urljoin==1.0.0
urllib3==2.2.2
virtualenv==20.26.3
webdriver-manager==4.0.1
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==6.4.post2
